Add files via upload
sshjail.sh is fully commented out, so every piece of the script has a description of what it does and why. Parsing.sh takes the passwd file on a linux  system and pipes it into a file called users.txt that can be exported or used for whatever other purpose.